Changed EnemyAI health values to floats, allowed player to shoot at enemy hull, reducted enemy difficulty scaling

+++ MonoGameWindowsStarter/Spaceship/Ship.cs
@@ -302,9 +302,9 @@ namespace MonoGameWindowsStarter.Spaceship
             //timer = new TimeSpan(0);
         }
 
-        public void AlterHealth(int damage)
+        public void AlterHealth(float damage)
         {
-            CurrentHealth -= damage;
+            CurrentHealth -= (int)damage;
                 
         }
